Rename commons directory to misc
