+++ lib/components/Icon.js
@@ -1,5 +1,5 @@
-import ns from "../commons/namespaces.js";
-import MetaData from "../commons/MetaData.js";
+import ns from "../misc/namespaces.js";
+import MetaData from "../misc/MetaData.js";
 import Pictogram from "./Pictogram.js";
 
 var warehouse = [];
