+++ lib/components/Sprite.js
@@ -1,5 +1,5 @@
-import { $, $$, $body, addClass, isInPage } from "../commons/utils.js";
-import ns from "../commons/namespaces.js";
+import { $, $$, $body, addClass, isInPage } from "../misc/utils.js";
+import ns from "../misc/namespaces.js";
 import Pictogram from "./Pictogram.js";
 
 var counter = 0;
