+++ lib/components/Theme.js
@@ -1,4 +1,4 @@
-import { $$, $body, isInPage } from "../commons/utils.js";
+import { $$, $body, isInPage } from "../misc/utils.js";
 import Icon from "./Icon";
 
 var warehouse = [];
