+++ lib/misc/MetaData.js
