+++ lib/misc/namespaces.js
