+++ lib/misc/utils.js
